changed address bus size from 32 bits to 64 bits

+++ include/buses.hpp
@@ -16,7 +16,7 @@ class Buses
 {
     private:
         uint64 dataBus;
-        uint32 addressBus;
+        uint64 addressBus;
         uint8 readWritePin;
         uint8 dataSizeBus;
         std::list<Peripheral *> observerList;
@@ -25,16 +25,16 @@ class Buses
         Buses();
         ~Buses();
         uint64 getDataBus(void);
-        uint32 getAddressBus(void);
+        uint64 getAddressBus(void);
         uint8 getReadWritePin(void);
         uint8 getDataSizeBus(void);
         void setDataBus(uint64 data);
-        void setAddressBus(uint32 address);
+        void setAddressBus(uint64 address);
         void setReadWritePin(uint8 pin);
         void setDataSizeBus(uint8 size);
         void addObserver(Peripheral *p);
-        void setBusesAndNotify(uint8 rw, uint32 addr, uint8 size);
-        void setBusesAndNotify(uint8 rw, uint32 addr, uint64 data, uint8 size);
+        void setBusesAndNotify(uint8 rw, uint64 addr, uint8 size);
+        void setBusesAndNotify(uint8 rw, uint64 addr, uint64 data, uint8 size);
         void notifyPeriph(void);
 };
